Fix #108 closures with params in REPL
